+++ liquidfun/Box2D/Testbed/Framework/ParticleParameter.cpp
@@ -20,16 +20,23 @@
 
 const ParticleParameter::Value ParticleParameter::k_particleTypes[] =
 {
-	{ b2_waterParticle, "water" },
-	{ b2_springParticle, "spring" },
-	{ b2_elasticParticle, "elastic" },
-	{ b2_viscousParticle, "viscous" },
-	{ b2_powderParticle, "powder" },
-	{ b2_tensileParticle, "tensile" },
-	{ b2_colorMixingParticle, "color mixing" },
-	{ b2_wallParticle, "wall" },
-	{ b2_barrierParticle | b2_wallParticle, "barrier" },
-	{ b2_staticPressureParticle, "static pressure" },
+	{ b2_waterParticle, ParticleParameter::k_DefaultOptions, "water" },
+	{ b2_waterParticle, ParticleParameter::k_DefaultOptions |
+				ParticleParameter::OptionStrictContacts, "water (strict)" },
+	{ b2_waterParticle, ParticleParameter::k_DefaultOptions |
+				ParticleParameter::OptionDrawAABBs, "water (bounding boxes)" },
+	{ b2_springParticle, ParticleParameter::k_DefaultOptions, "spring" },
+	{ b2_elasticParticle, ParticleParameter::k_DefaultOptions, "elastic" },
+	{ b2_viscousParticle, ParticleParameter::k_DefaultOptions, "viscous" },
+	{ b2_powderParticle, ParticleParameter::k_DefaultOptions, "powder" },
+	{ b2_tensileParticle, ParticleParameter::k_DefaultOptions, "tensile" },
+	{ b2_colorMixingParticle, ParticleParameter::k_DefaultOptions,
+		"color mixing" },
+	{ b2_wallParticle, ParticleParameter::k_DefaultOptions, "wall" },
+	{ b2_barrierParticle | b2_wallParticle, ParticleParameter::k_DefaultOptions,
+		"barrier" },
+	{ b2_staticPressureParticle, ParticleParameter::k_DefaultOptions,
+		"static pressure" },
 };
 const ParticleParameter::Value *ParticleParameter::k_particleTypesPtr =
 	ParticleParameter::k_particleTypes;